requirements.txt
pydantic==2.7.2
pyyaml
ruff

